+++ WebServer/Domain/DepositHistory.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
 
 namespace HamsterCoin.Domain
 {
@@ -11,7 +10,7 @@ namespace HamsterCoin.Domain
         [Key]
         [Column("id")]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [ForeignKey(nameof(User))]
         [Column("user_id")]
