+++ WebServer/Domain/User.cs
@@ -1,6 +1,5 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using Microsoft.EntityFrameworkCore;
 
 namespace HamsterCoin.Domain
 {
@@ -12,10 +11,10 @@ namespace HamsterCoin.Domain
         [Key]
         [Column("id")]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Required]
-        [Column("mail")] // Переглянути можливість [EmailAddres]
+        [Column("mail")]
         [EmailAddress]
         [MaxLength(100)]
         public string Mail { get; set; }
